Stabilize SDK validation and CI (#2)
* fix(schema): enforce canonical event ID format

* fix(python): resolve shared fixtures inside sdists

* fix(python): make packaged schema generation work

* fix(schema): reject impossible webhook versions

* fix(node): reject empty canonical headers

* fix: normalize exported helper errors

* fix(go): accept raw JSON in ParseWebhookEvent

* fix(go): use fixed-width JSON integer fields

* fix(go): keep ParseWebhookEvent string inputs strict

* chore: restore clean docs and static checks

Align the repository docs with the actual layout and keep the SDKs lint- and typecheck-clean so the repo stays release-ready.

* fix(python): make model generation reproducible

Patch the generated models after datamodel-codegen runs so regeneration preserves the SDK-specific validation and serialization behavior required by release-check.

* chore: add a local ci target

Mirror the repo checks behind one command and keep Python coverage output from dirtying the worktree.

* chore: lower supported SDK minimum versions

* fix: align SDK schema serialization and null handling

Keep generated Python JSON dumps consistent with model_dump, reject parsed JSON null before Go webhook reflection, and document the actual generated Go schema artifact.

* chore(node): raise minimum supported Node version to 20

Update the declared Node support floor to 20 and expand the CI matrix to cover Node 20, 22, and 24 so supported and tested versions stay aligned.

---------

+++ sdk-go/types.go
@@ -150,7 +150,7 @@ func (e EmailReceivedEvent) GetEvent() string { return e.Event }
 
 type Delivery struct {
 	EndpointID  string `json:"endpoint_id"`
-	Attempt     int    `json:"attempt"`
+	Attempt     int64  `json:"attempt"`
 	AttemptedAt string `json:"attempted_at"`
 }
 
@@ -193,8 +193,8 @@ type RawContent struct {
 	Included       bool    `json:"included"`
 	Encoding       *string `json:"encoding,omitempty"`
 	ReasonCode     *string `json:"reason_code,omitempty"`
-	MaxInlineBytes int     `json:"max_inline_bytes"`
-	SizeBytes      int     `json:"size_bytes"`
+	MaxInlineBytes int64   `json:"max_inline_bytes"`
+	SizeBytes      int64   `json:"size_bytes"`
 	SHA256         string  `json:"sha256"`
 	Data           *string `json:"data,omitempty"`
 }
@@ -221,9 +221,9 @@ type EmailAddress struct {
 type WebhookAttachment struct {
 	Filename    *string `json:"filename"`
 	ContentType string  `json:"content_type"`
-	SizeBytes   int     `json:"size_bytes"`
+	SizeBytes   int64   `json:"size_bytes"`
 	SHA256      string  `json:"sha256"`
-	PartIndex   int     `json:"part_index"`
+	PartIndex   int64   `json:"part_index"`
 	TarPath     string  `json:"tar_path"`
 }
 
@@ -245,9 +245,9 @@ type SpamAssassinAnalysis struct {
 type ForwardAnalysis struct {
 	Detected            bool            `json:"detected"`
 	Results             []ForwardResult `json:"results"`
-	AttachmentsFound    int             `json:"attachments_found"`
-	AttachmentsAnalyzed int             `json:"attachments_analyzed"`
-	AttachmentsLimit    *int            `json:"attachments_limit"`
+	AttachmentsFound    int64           `json:"attachments_found"`
+	AttachmentsAnalyzed int64           `json:"attachments_analyzed"`
+	AttachmentsLimit    *int64          `json:"attachments_limit"`
 }
 
 type ForwardResult struct {
@@ -307,7 +307,7 @@ type DKIMSignature struct {
 	Selector *string    `json:"selector,omitempty"`
 	Result   DkimResult `json:"result"`
 	Aligned  bool       `json:"aligned"`
-	KeyBits  *int       `json:"keyBits,omitempty"`
+	KeyBits  *int64     `json:"keyBits,omitempty"`
 	Algo     *string    `json:"algo,omitempty"`
 }
